+++ Assets/Scripts/GunZ/GunSO.cs
@@ -19,7 +19,7 @@ public class GunSO : ScriptableObject
     public int bulletsPerClick;
     public int damage;
     public int critChance;
-    public int critMultiplier;
+    public float critMultiplier;
     public int hitChance;
     public int chanceToHitOtherParts;
     public int attackRange;
